issue-033 메인 프래그먼트 수정
