+++ app/src/main/AndroidManifest.xml
@@ -10,22 +10,18 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
 
-        <activity android:name=".view.LogInActivity">
-        </activity>
-        <activity android:name=".view.SplashActivity">
-
-        </activity>
-
+        <activity android:name=".view.main.SplashActivity"/>
+        <activity android:name=".view.main.LogInActivity"/>
         <activity
             android:name=".view.main.MainActivity"
             android:label="@string/title_activity_main">
-
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".view.upload.UploadActivity" />
 
     </application>
