+++ app/src/main/java/com/boostcamp/dreampicker/view/main/LogInActivity.java
@@ -1,4 +1,4 @@
-package com.boostcamp.dreampicker.view;
+package com.boostcamp.dreampicker.view.main;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -8,18 +8,20 @@ import android.os.Bundle;
 import com.boostcamp.dreampicker.R;
 import com.boostcamp.dreampicker.view.main.MainActivity;
 
-public class SplashActivity extends AppCompatActivity {
+public class LogInActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_splash);
+        setContentView(R.layout.activity_log_in);
 
         try{
             Thread.sleep(500);
         }catch (InterruptedException e){
             e.printStackTrace();
         }
-
+        Intent intent = new Intent(getApplicationContext(), MainActivity.class);
+        startActivity(intent);
+        finish();
     }
 }
