+++ app/src/main/java/com/boostcamp/dreampicker/view/profile/ProfileFragment.java
@@ -20,7 +20,7 @@ public class ProfileFragment extends BaseFragment<FragmentProfileBinding> {
 
     }
 
-    public static ProfileFragment getInstance(){
+    public static ProfileFragment newInstance(){
 
         return new ProfileFragment();
     }
