+++ app/src/main/res/layout/activity_log_in.xml
@@ -5,7 +5,7 @@
     <RelativeLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".view.LogInActivity">
+        tools:context=".view.main.LogInActivity">
 
 
         <ImageView
